Fixing volatile shared memory access and updated the stress-test
We should be using volatile keyword for the array, but it only seems to be working with built-in types. Therefore added syncthreads()-calls to appropriated places.

Also added an option for stress-test- that will only check that the checksum of the original histogram is correct (instead of checking each bin separately and comparing against cpu-results)