Added functionality to perform mappings and reductions.
The APIs are very similar to those of the Histogram API and should be fairly simple to use. Also added a random-number generator
by David B. Thomas of Imperial College under the BSD License - the rnd-generator seems high quality and fast, and most importantly
generates random numbers on the fly on the GPU with little overhead. The Ising-model example makes extensive use of this RND
generator